+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.0.3-beta.0] - 2025-09-11
+
+- Dynamic dropdowns for Video ID, Image ID, Audio ID and Collection ID wherever needed
+- Addressed bugs related to duplicate ids in List, Get and Delete Scene Indexes
+- Code cleanup of commented out code
+
+## [1.0.2]() - 2025-09-7
+
+## Removed
+
+- Removed Get, List and Connect RTStream operations
+- Remove Make Collection Private and Make Collection Public operations
+
 ## [1.0.1]() - 2025-09-05
 
 ### Changed